services/api/pyproject.toml
[project]
name = "starvit-api"
version = "0.1.0"
description = "Starvit Backend API"
authors = [
    {name = "Admin", email = "admin@starvit.ca"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi>=0.125.0,<0.126.0",
    "uvicorn>=0.38.0,<0.39.0",
    "httpx>=0.28.1,<0.29.0",
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "python-dotenv>=1.2.1,<2.0.0",
    "pyjwt[crypto]>=2.10.1,<3.0.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "poetry-core (>=2.0.0,<3.0.0)"]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100
select = ["E", "F", "I", "W"]
ignore = []
